chore: fix date and event enter
